requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.13
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
